Renamed tests folder to integration_tests Since tests are actually endpoint tests, they cover more than unit testing, as they require integration between the routers and the core service layer, who also require integration with the port layer
