+++ tests/integration_tests/test_result.py
@@ -10,7 +10,7 @@ from core.model.result import ResultCreation
 
 from tests.mocks.mongo_mock import MongoMock
 
-from tests.unit_tests.config import (
+from tests.integration_tests.config import (
     configure_ports_without_auth,
     configure_ports_with_auth,
 )
